Choose num of words to split on based on text area
This has been glitchy when screen is on landscape so instead of using
view width and height, we use the area of the textbox element in order
to determine at how many words we should split the text. This still
leaves a bit of an issue when the browser font-size is changed by the
user but it's proven to be a bit of an issue to solve all these problems
well.